+++ wasm/OpenSimplexNoise.cpp
@@ -1,10 +1,10 @@
 #include "OpenSimplexNoise.h"
 
-const double OpenSimplexNoise::STRETCH_3D = -1.0 / 6.0;
-const double OpenSimplexNoise::SQUISH_3D = 1.0 / 3.0;
-const double OpenSimplexNoise::NORM_3D = 1.0 / 103.0;
+const float OpenSimplexNoise::STRETCH_3D = -1.0 / 6.0;
+const float OpenSimplexNoise::SQUISH_3D = 1.0 / 3.0;
+const float OpenSimplexNoise::NORM_3D = 1.0 / 103.0;
 
-std::array<double, 72> OpenSimplexNoise::gradients3D;
+std::array<float, 72> OpenSimplexNoise::gradients3D;
 
 std::vector<OpenSimplexNoise::Contribution3*> OpenSimplexNoise::lookup3D;
